fix(wallets): close the defects a Codex and CodeRabbit pass found
Nothing here changes the design. Each item is a place where the code did
not do what its own docstring, or the money, required.

A stream entry left pending was never redelivered. Both workers'
docstrings promised that a retryable failure leaves the entry "pending for
normal consumer-group redelivery", but neither enabled the consumer's
reclaim pass, and the read path only ever asks Redis for new entries. So
"leave it pending" meant "drop the charge silently". Both now opt in.
Redelivery is safe because the measurement insert is idempotent on its
measurement id and settlement is idempotent on the idempotency key. Only
an entry that cannot be decoded at all is treated as permanently failed:
money must not be dropped because the write path was down.

A month-end anchor selected a period that had already ended. The billing
window compared today against the raw anchor day but clamped it when
building the dates, so anchor 31 on 28 February produced a window ending
before now, and proration of the remainder returned zero.

A plan change prorated both sides over one window, and read the anchor
after the event had overwritten it. Cancellation therefore clawed back
nearly the whole allowance whatever the customer had used. There are two
windows: the outgoing allowance can only be clawed out of the period it
was granted for, and the incoming allowance covers the period the
subscription now carries, which on a new subscription is Stripe's own
billing cycle anchor. They are identical when the anchor does not move.

Two metric columns were 32-bit while every wallet money column is 64-bit,
so a component above about 2147 dollars failed to persist. Both are
widened in the same unreleased migration.

Two base classes were named WalletError, one per module, so catching the
exported one missed exactly the errors that mean a charge did not settle.

An envelope whose JSON parsed to something other than an object died later
on attribute access, which no worker classified as terminal, so it retried
forever. It is now rejected as malformed.

A retried signup sized the wallet floor from the default plan rather than
the plan the organization was actually on.

The wallet test fakes did not scope credits to an organization, so they
accepted selections the real DAO rejects.

The backfill migration read every organization in one statement. It now
pages by key.

+++ api/ee/databases/postgres/migrations/tracing_ee/versions/ee0000000002_add_measurements.py
@@ -84,8 +84,12 @@ def upgrade() -> None:
         sa.Column("id", sa.UUID(), nullable=False),
         sa.Column("measurement_id", sa.UUID(), nullable=False),
         sa.Column("key", sa.VARCHAR(), nullable=False),
-        sa.Column("value", sa.Integer(), nullable=False),
-        sa.Column("cost_musd", sa.Integer(), nullable=True),
+        # Both metric columns are 64-bit: `cost_musd` is money and a millionth-of-a-dollar
+        # unit overflows a 32-bit column at ~2147 US dollars, and `value` is an unbounded
+        # count from the envelope. A row the column rejects is retried forever by the
+        # measurement worker, not dropped.
+        sa.Column("value", sa.BigInteger(), nullable=False),
+        sa.Column("cost_musd", sa.BigInteger(), nullable=True),
         sa.Column(
             "created_at",
             sa.TIMESTAMP(timezone=True),
